Implemented orders by user call

+++ src/main/java/it/unibz/infosec/examproject/order/domain/ManageOrders.java
@@ -9,6 +9,7 @@ import it.unibz.infosec.examproject.util.crypto.rsa.RSA;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -26,13 +27,10 @@ public class ManageOrders {
         this.manageProducts = manageProducts;
     }
 
-    private Order validateOrder (Long id) {
-
-        Optional<Order> maybeOrder = orderRepository.findById(id);
-
-        if(maybeOrder.isEmpty())
+    private Order validateOrder(Long id) {
+        final Optional<Order> maybeOrder = orderRepository.findById(id);
+        if (maybeOrder.isEmpty())
             throw new IllegalArgumentException("Order with id '" + id + "' does not exist yet!");
-
         return maybeOrder.get();
     }
 
@@ -54,10 +52,10 @@ public class ManageOrders {
     }
 
     public Order approveOrder(Long idOrder) throws Exception {
-        Order order = readOrder(idOrder);
+        final Order order = readOrder(idOrder);
         boolean orderValidity = isSignatureOrderValid(order.getId());
 
-        if(!orderValidity)
+        if (!orderValidity)
             throw new Exception("Order digital signature is not valid!");
 
         order.setApproved(true);
@@ -65,30 +63,29 @@ public class ManageOrders {
     }
 
     public boolean isSignatureOrderValid(Long idOrder) {
-        Order order = readOrder(idOrder);
-        UserEntity client = manageUsers.readUser(order.getClientId());
+        final Order order = readOrder(idOrder);
+        final UserEntity client = manageUsers.readUser(order.getClientId());
         byte[] DSA = order.getDSA();
-        String retrievedDigest = RSA.decryptToString(DSA, client.getPublicKey(), client.getNKey());
-        String computedDigest = Hashing.getDigest(order.getOrderDocument());
+        final String retrievedDigest = RSA.decryptToString(DSA, client.getPublicKey(), client.getNKey());
+        final String computedDigest = Hashing.getDigest(order.getOrderDocument());
 
         return retrievedDigest.equals(computedDigest);
     }
 
     public Order readOrder(Long id) {
-
-        Order order = validateOrder(id);
-
-        return order;
-
+        return validateOrder(id);
     }
 
-    public Order deleteOrder (Long id) {
-
-        Order order = validateOrder(id);
-
+    public Order deleteOrder(Long id, Long customer) {
+        final Order order = validateOrder(id);
+        if (!order.getClientId().equals(customer)) {
+            throw new IllegalArgumentException("Only the holder of an order can delete it");
+        }
         orderRepository.delete(order);
-
         return order;
     }
 
+    public List<Order> getByCustomer(Long customerId) {
+        return orderRepository.findByClientId(customerId);
+    }
 }
